Propagate the register size through field values
PWM registers are 16 bits. But, the imxrtral.py script is generating
masks, offsets, and enumerated values that are hard-coded to u32s. When
we try to use bit operations -- like those inlined by read_reg!,
write_reg!, or modify_reg! -- the compiler yells at us. It's because
we're trying to do bit operations on incompatible types: u32 and u16.

The masks, offsets, and enumerated values for a field should be based
on the overarching register. This commit updates the RAL generation
script to propagate the register size to the nodes that generate the
corresponding Rust code. We should see that the constants associated
with PWM fields are now correctly represented as u16s.

After applying this commit, you should re-generate the RAL. Otherwise,
you may find that the PWM HAL modules will not build.

This commit should be ported to the stm32ral project.