Remove some leftover STM32 references
I missed some comments in the RAL like "Set PA3 high," and "Reset the
JTAG pins." These are specific to STM32 MCUs. I made those comments
more general for the i.MX RT (we're doing far-less interesting
field manipulations).

Also, fix one of the manually-expanded examples that seemed to be a
copy-and-paste error from the stm32ral documentation. Looks like
we were shifting and masking a field by the wrong mask and offset.