Conditionally emit __INTERRUPT link section
I came upon this when I was writing automated unit tests for the
imxrt-uart-log logging crate. I wanted to 'cargo test' on my
mac. But, I hit an error:

  LLVM ERROR: Global variable '__INTERRUPTS' has an invalid section
  specifier '.vector_table.interrupts': mach-o section specifier
  requires a segment whose length is between 1 and 16 characters.

Rather than always emitting the link section for __INTERRUPTS, we
conditionally emit the link section when building for ARM targets.
This lets me run unit tests on my host, while still including the
link section for the targets that need it.