Patch RIDMAE in 1015, 1021 SVDs
Not sure why this field is missing. The field is valid across all
BAUD registers, for all considered variants. We found this when
discussing a split HAL.

imxrtral.py still doesn't correctly fold the LPUART peripherals
together across all devices. It's either because

- it treats the missing enumerated values as semantic, incompatible
  differences
- it's not ordering the fields by bit offset, so it treats them as
  difference (observed by comparing the generated code)

Might need to investigate further if we go with this approach.