Correct USBCMD[ATDTW] bit offset
The bit is correct in the 1011 and 1015 SVDs. But, it's incorrect
in the 1021, 1051, 1052, 1061, 1062, and 1064 SVDs. I confirmed
that bit offset 14 is correct by studying the 1020, 1050, and 1060
reference manuals (skipped the 1064 RM, assumed it's consistent).

Note that this patch will affect the 1011 and 1015 SVDs, since it's
included by default. But, it will have no effect.

Tried to source an updated 1062 SVD file, thinking it could have been
fixed upstream. But, I didn't find an updated SVD from developer.arm.
com. Didn't consider any other chip SVDs.