Repatch USB peripherals for module consistency
When developing the imxrt-usbd driver, I hit an edge in the
peripheral module names. Chips that only have 1 USB peripheral had
a module name imxrt_ral::usb1, and chips that have 2 USB peripherals
had a module name imxrt_ral::usb. The inconsistency makes it hard
to develop a driver to the RAL that doesn't require extraneous
features.

This commit undoes the USB -> USB1 rename. The end result is that
the USB's Instance type is always accessible at imxrt_ral::usb.
With this in place, we can design a USB driver that doesn't require
its own feature flags to correctly identify the path.

This is a breaking change for all 1010, 1015, and 1020 chip users,
so it'll be released as version 0.5.