Always generate instances with typenums
Before this commit, we would conditionally generate an Instance with
a typenum. Now, we always do it, even if the peripheral doesn't
support multiple instances. We do this to help others create more
general code.

Consider the following:

	use imxrt_ral::usb;

	fn take_usb<N>(_: usb::Instance<N>) { /* ... *}

The above assumes that the USB instance is tagged with an instance
identifier. With the previous implementation, this is true for 1050
and 1060 chips, but not for 1010 chips. A user who wanted to support
all chips would need to do something like

	use imxrt_ral::usb

	#[cfg(feature = "1010")]
	fn take_usb(_: usb::Instance) { /* ... */ }

	#[cfg(any(feature = "1050", feature = "1060"))]
	fn take_usb<N>(_: usb::Instance<N>) { /* ... */ }

which is cumbersome, and requires the user to provide their own
features.

By always generating generic instances, the user doesn't have to use
conditional compilation to choose a take_usb implementation. The same
function, generic over N, will work for all chips. See the compile
tests in the 'usage' documentation.

Inline a "SingleInstance" instance identifier, chosen to be U0, to
make codegen simpler.