Associate interrupt signals with instances
All instances now have a method, interrupts(), which returns zero or
more interrupt signals. We derive the signals from the SVD. Peripheral
instance modules also expose a constant array of interrupt signals.

We need to create a stub interrupt type to support documentation
generation. Without it, there would either be symbol collisions when
re-exporting the Interrupt enum, or we'd have to change how peripheral
and instance code is shared, and how we're referencing interrupts. It
seems simpler to create the stub, and to qualify how the stub resolves
when building for a specific chip.