Remove NVIC, SCB; move interrupts to peripherals
Following #7, this commit removes the NVIC and SCB peripherals from
imxrt-ral. These two peripherals are listed as i.MX RT-specific
peripherals in the SVDs. We delete the peripherals from the SVDs.
This has the nice outcome of removing many "Reserved" symbols in
the RAL.

The NVIC was including the IOMUXC_GPR and XBAR interrupts. We add
patches that associate the interrupt signals with the peripherals,
which is what nearly all other peripherals already do. The changes
are backwards compatible; all interrupt signals and weak symbols are
the same.

Looks like the 1050 and 1060 SVDs have an undocumented JTAG interrupt
signal. Signal 54 is listed as "reserved" in the reference manuals.
This commit removes that symbol. It also marks as reserved the CM7
interrupt signals, since they were tied to one of the ARM core
peripherals.