Add usage documentation
Describes how to use the RAL API, and design higher-level APIs using
the RAL. Implemented as a separate markdown file so it could be studied
within the repository. It also makes it easier to maintain within the
library documentation (we don't need to inline it within the Python
script or anything). Documentation examples are tested as if they were
written inline.