Always generate instance types, implement Valid
There's no harm in defining the Instance type in a 'nosync' build, since
there's no way to realize an instance.  The array of interrupts is tied
to the instance type, even if 'nosync' is enabled. As of this commit,
the test for associated interrupts pass.

Additionally, always implement Valid for instances, even if 'nosync' is
enabled.