Ensure alignment when computing reserved registers
If the routine for computing reserved peripheral memory places at least
one u32 after a u16 register, the compiler could insert an extra u16 of
padding to align the u32 reservation. The extra padding is not
considered by the routine. This can result in incorrect register offsets
for all registers placed after the reservation.

This commit ensures that we only use larger-than-byte padding when the
address is aligned for that primitive. If the address isn't aligned, use
the next-largest primitive. Byte array reservations should be the safest
approach, but going that route changes all _reserved members of all
register blocks, making it harder to see what peripherals were affected
by this issue.

I wasn't having much luck with a FlexPWM driver on the 1011 until this
commit. This commit also adds a local test to ensure proper reservations
and padding; the test now passes.