Add preliminary support for 1176 cores
The RAL builds with two new features for the two 1176 cores. There's
some questionable code generation that should be addressed. Quick
observations are documented below. The "cm4" and "cm7" suffixes come
from the two SVDs.

Peripheral consolidation can be too aggressive. All GPIOs are available
in the cm7_gpio module. Not only is it poorly named w.r.t. the CM4, but
it allows the CM4 to access the CM7's dedicated GPIOs. I don't think the
hardware permits this. (The AIPS1 is drawn so that these GPIOs are
reachable through the performance crossbar, but I think that line is
supposed to touch the CM7's FlexRAM, not the GPIOs...) We should ensure
that the GPIOs and CM_GPIO[2|3] are not consolidated, and this is
handled as edge cases directly in the script.

Other edge cases:

- Keep combining the XECC peripherals, but manually number the SEMC
  variant with 0.
- Manually number the PGMC instances, starting at 1. This means that the
  instance numbers have no bearing on these peripherals. This numbering
  scheme for these peripherals shouldn't be relied upon.

(As I wrote this, I realized that the CM4 SVD actually contains the
CM7_GPIO register blocks. I'm guessing that's an SVD defect. The CM4's
register block definitions reach into CM4 reserved memory... Maybe
fixing the SVD will be enough.)

Similarly, the ETH, Video, and Audio PLLs are consolidated in an
"audio_pll" module. It's not obvious that ETH and Video PLLs are in that
module. Nevertheless, I'm removing the peripherals; they're aliased, and
all pointing to address zero, which isn't right.

These SVDs have a bunch of duplicated enumeratedValues that cause
duplicate definitions in the RAL. Nearly all of the SVD patches are
designed to remove or simplify these enums.