Add support for non-u32 bitfields
The BitSize newtype helped with refactoring, and could be removed if
deemed too noisy. After this commit, the size of a set of fields is
based on the size of the register.

Before this commit, the registers for an MIMXRT PWM peripheral were all
incorrectly u32s. After this commit, I observed that code generated for
an MIMXRT PWM peripheral used u16 throughout. It still might use u8 for
individual fields, but that doesn't affect the MMIO register
interactions. Only tested by extracting peripherals, then generating
blocks based on the YAML output; haven't tested a full SVD -> PAC
transform, but I would expect that to work.