Favor an svdtools tool written in Rust
There's both Rust- and Python-based svdtools available. We've been using
the Python one because it existed well before the Rust one. But now that
we don't have a Python-based RAL code generator, there's less reason to
keep other Python tooling.

This commit adopts the Rust svdtools binary. It supports the same
workflows and YAML as the Python tool. The small code changes are due to
issue 110 in svdtools, which was fixed in the Rust tool but not the
Python tool. As of this commit, we don't need any Python tooling to
generate the RAL. (There's still some remnants of stm32ral Python
scripts in the Makefile and repo, but they're not on our common workflow
path.)