Add number functions for ptr -> N runtime query
If you discard type information, it's nice to have a way to re-acquire
the N associated with a peripheral. This commit introduces the `number`
functions for that purpose. This can let users create thin drivers just
around a register block static reference, and only acquire the
number at runtime.

Document this new function as an advanced usage, along with one other
(unrelated) usage pattern.