Revise tests
Delete associated_interrupts, since we're no longer generating this API.
My experiment that used these never really went anywhere. Might revisit
this later.

Delete the take_release tests; there's no more take / release.

Add a test that asserts the CCM peripheral for the 1170 looks OK. This
peripheral is one of the more complex peripherals that we generate from
SVD; it has clusters and register arrays. It's helpful to test the
codegen with an advanced register block.