Expose an Instances aggregate
It's the same as RTIC's Peripherals, but supported for all users. I'm
dogfooding this idea in a BSP design; it simplifes BSP setup for both
the user and the BSP author.

I'm now regretting 0b97844b89, so this commit undoes that commit.
Suppose each core of an 1170 is to run an RTIC application, and each
application is separately compiled and linked. If the user enables the
'peripherals = true' feature of the RTIC app macro, then each
application's init function observes duplicate chip peripheral
instances, and each application may think it has exclusive ownership of
these instances. We're trying to make the user think about this, so
letting RTIC hide this behavior might not be the best approach.

By removing `Peripherals`, we prevent RTIC's macro from auto-allocating
all peripheral instances without an `unsafe` block. Based on their own
specific needs, users can opt-in to RTIC support by defining a module
with something like

  pub type Peripherals(pub imxrt_ral::Instances);

  impl Peripherals {
    pub const unsafe fn steal() -> Self {
      Self(imxrt_ral::Instances::instances())
    }
  }

and re-exporting the other RTIC-specific things (interrupt, Interrupt,
NVIC_PRIO_BITS, etc.). If you take this approach in a BSP, instruct
users to use your BSP / crate for the RTIC application.