Require field name equality for field equivalence
Omitting this check allowed the combine pass to eagerly merge fieldsets
it believed were equivalent. Consider the LPI2C.SIER register. On the
1000 series, the "address match 1 interrupt enable" field is called
"AM1IE." And on the 1100 series, the same field is called "AM1F." Same
meaning, but a different name. By eagerly combining, we have no need to
generate another lpi2c.rs files, and we have 1675 fewer lines of code.

The assumption fails when that name difference is meaningful, like when
NXP changes the CCM.CS1CDR field at offset 25 from FLEXIO1_CLK_PODF to
FLEXIO2_CLK_PODF. Those aren't equivalent fields, but we don't know that
because we're not checking the field name.

It's always correct to not combine. And turns out it's not adding
too much code; we can patch the small name corrections like the LPI2C
example. Those patch commits are children of this commit.