Rename LPI2C.SIER AM1IE field on 1100 series
See the parent commit for more context. With this patch, we can still
combine the LPI2C peripherals across all peripherals.

Since field combinations wasn't considering names, the combiner was free
to select any version of the field as the base. It happened to pick the
"less good" field name for all peripherals. This commit ends up changing
the source as a consequence of the rename.