Refactor equivalence checks in raltool
A function pointer becomes a trait, enabling the equivalence check
contexts to hold non-static data. Equivalence implementations are now
supplied the path of the IR element they're evaluating. As of this
commit, there's no change to the combiner's behavior.

The check for equivalent enums can now check for variant names. This
will depend on a configuration, which I'll introduce in another commit.