Add raltool config for strict enum descriptions
Similar to strict enum names, this configuration instructs the combiner
to compare enum variant descriptions, the human-readable strings
associated with items. If they're not equal, then the enums cannot be
combined.

When applied to IOMUXC peripherals, this ensures that the documentation
associated with each enum variant is correct. The next commit checks in
the updates.