Fix FlexIO pins, register counts for 1060 MCUs
The 1060 SVDs suggest that FlexIO2 and 3 have 16 pins. They should have
32 pins. This commit updates select field widths to five bits to express
the additional pins. This change affects FlexIO1, and downstream users
will need to handle the fact that these extra pins are invalid.
The committer cross-checked all increased field widths against the 1060
RM.

This commit also increases the timer and shifter count for the FlexIO
peripherals. This implementation has eight timers and shifters, not
four.