Exclude interrupt vector table from host build
I need a way to exclude the vector table from being linked into the
build script.  Was going to try target_arch = "arm" and realized that
it's possible for the host arch to be "arm" too.

Since the firmware is built with target_os = "none" and the build script
is built on a non-bare-metal system, I went with this as the extra
filter. It seems to be working.