Remove NVIC, SCB from the 1176
The 1176 imxrt-ral modules no longer provide an API for the NVIC or SCB.
It's the same process as b1fee46, but applied to the 1176 modules. I
re-introduced core interrupts by attaching them to the miscellaneous
control module (MCM). All separate GPC peripheral blocks are logically
within a single 'GPC' space, so selecting the standby sub-module was an
arbitrary decision.

Notice that this commit marks the GPU2D interrupt as reserved. This is
intentional; the SVDs do not describe a GPU2D peripheral, so imxrt-ral
doesn't have a corresponding module. This interrupt probably
piggy-packed on the NVIC. We can re-introduce that interrupt when we
describe the GPU2D peripheral.