Force common blocks across MCUs
A `ForceCommonBlock` directive tells the combiner that it's safe to
treat a specific MCU's peripheral block as common across all MCUs. It's
an override for the combiner's default "all registers, all fields, all
enums" evaluation.

We can use this configuration to change how we codegen the RAL for the
1180. I'm focusing on the peripherals that have driver support in
imxrt-hal.

I'm using SVD patches to inflence how raltool selects its root blocks in
its IR. I'm not sure why we're losing some documentation comments, but
my guess is that it's happening during that rebase.