Patch 1180 USB peripherals for consolidation
I'm applying the 1170 patches upon the 1180. With this in place, raltool
can see that the USB peripheral block is equivalent to all the other USB
peripheral blocks. Caveat: I'm only assuming they're equivalent, and I
haven't tested this on hardware.

I made sure to derive values from the 1180 SVDs when re-deriving the
second peripheral block instances. We can show this by making sure the
block addresses in the Rust code are unchanged.