Generate instances without numbering requirement
We can still emit the `number` function if there's just one instance.
We've been doing this for sole instance types, so this maintains
backwards compatibility.

If there are multiple peripherals associated with the same block, and if
they're not numbered, then we can treat them as distinct. We abuse the
const generics to stamp out unique type aliases, and we start counting
backwards from 255. We signal through documentation that this const
generic is meaningless.

There's other ways to do this; one way is to introduce a unique type
state to the generic `Instance`. By default, use `()`. Then let
implementations define their own type state when they specify their
instance aliases, thus generating unique types with the same "sole
instance" const generic. It's more code to take this approach, and I'm
time limited.